refactor: extract hook logic from BoardsWrite component
